Simplify indexer module
Promote services/search_index.service to services/search_index.py

- Inline search_index service factory
- Inline search_index._queue into the service
- Simplify search_index module
