--- h/services/search_index/__init__.py
@@ -1,2 +1,0 @@
-from h.services.search_index.service import SearchIndexService
-from h.services.search_index.service_factory import factory
