+++ tests/unit/h/search/conftest.py
@@ -2,9 +2,9 @@ from unittest import mock
 
 import pytest
 
+from h.search.index import BatchIndexer
 from h.services.group import GroupService
 from h.services.search_index import SearchIndexService
-from h.services.search_index._queue import Queue
 
 
 @pytest.fixture
@@ -47,15 +47,20 @@ def index_annotations(es_client, search_index):
 
 @pytest.fixture
 def search_index(  # pylint:disable=unused-argument
-    es_client, pyramid_request, moderation_service, annotation_read_service
+    es_client,
+    pyramid_request,
+    moderation_service,
+    annotation_read_service,
+    queue_service,
 ):
     return SearchIndexService(
         pyramid_request,
-        es_client,
-        session=pyramid_request.db,
+        es=es_client,
+        db=pyramid_request.db,
         settings={},
-        queue=mock.create_autospec(Queue, spec_set=True, instance=True),
         annotation_read_service=annotation_read_service,
+        queue_service=queue_service,
+        batch_indexer=mock.create_autospec(BatchIndexer, spec_set=True, instance=True),
     )
